+++ src/client/transport/local.test.js
@@ -10,7 +10,7 @@ import { createStore } from 'redux';
 import { Local, LocalMaster } from './local';
 import Game from '../../core/game';
 import { makeMove, gameEvent } from '../../core/action-creators';
-import { CreateGameReducer } from '../../core/reducer';
+import { InitializeGame, CreateGameReducer } from '../../core/reducer';
 
 describe('LocalMaster', () => {
   const game = Game({});
@@ -91,7 +91,9 @@ describe('Local', () => {
     let store = null;
 
     beforeEach(() => {
-      m.store = store = createStore(CreateGameReducer({ game }));
+      const reducer = CreateGameReducer({ game });
+      const initialState = InitializeGame({ game });
+      m.store = store = createStore(reducer, initialState);
     });
 
     test('returns a valid store', () => {