+++ src/core/reducer.js
@@ -10,7 +10,7 @@ import { parse, stringify } from 'flatted';
 import * as Actions from './action-types';
 import { Random } from './random';
 import { Events } from './events';
-import * as plugins from './plugins/main';
+import * as plugins from '../plugins/main';
 
 /**
  * Moves can return this when they want to indicate
@@ -102,19 +102,15 @@ export class ContextEnhancer {
 }
 
 /**
- * CreateGameReducer
+ * InitializeGame
  *
- * Creates the main game state reducer.
+ * Creates the initial game state.
+ *
  * @param {...object} game - Return value of Game().
  * @param {...object} numPlayers - The number of players.
  * @param {...object} multiplayer - Set to true if we are in a multiplayer client.
  */
-export function CreateGameReducer({
-  game,
-  numPlayers,
-  multiplayer,
-  setupData,
-}) {
+export function InitializeGame({ game, numPlayers, setupData }) {
   if (!numPlayers) {
     numPlayers = 2;
   }
@@ -128,7 +124,7 @@ export function CreateGameReducer({
   ctx._random = { seed };
 
   // Pass ctx through all the plugins that want to modify it.
-  ctx = plugins.SetupCtx(ctx, game.plugins);
+  ctx = plugins.ctx.setup(ctx, game);
 
   // Augment ctx with the enhancers (TODO: move these into plugins).
   const apiCtx = new ContextEnhancer(ctx, game, ctx.currentPlayer);
@@ -137,7 +133,7 @@ export function CreateGameReducer({
   let initialG = game.setup(ctxWithAPI, setupData);
 
   // Pass G through all the plugins that want to modify it.
-  initialG = plugins.SetupG(initialG, ctxWithAPI, game.plugins);
+  initialG = plugins.G.setup(initialG, ctxWithAPI, game);
 
   const initial = {
     // User managed state.
@@ -159,6 +155,9 @@ export function CreateGameReducer({
 
     // A snapshot of this object so that actions can be
     // replayed over it to view old snapshots.
+    // TODO: This will no longer be necessary once the
+    // log stops replaying actions (but reads the actual
+    // game states instead).
     _initial: {},
   };
 
@@ -172,6 +171,18 @@ export function CreateGameReducer({
   const deepCopy = obj => parse(stringify(obj));
   initial._initial = deepCopy(initial);
 
+  return initial;
+}
+
+/**
+ * CreateGameReducer
+ *
+ * Creates the main game state reducer.
+ * @param {...object} game - Return value of Game().
+ * @param {...object} numPlayers - The number of players.
+ * @param {...object} multiplayer - Set to true if we are in a multiplayer client.
+ */
+export function CreateGameReducer({ game, multiplayer }) {
   /**
    * GameReducer
    *
@@ -179,7 +190,7 @@ export function CreateGameReducer({
    * @param {object} state - The state before the action.
    * @param {object} action - A Redux action.
    */
-  return (state = initial, action) => {
+  return (state = null, action) => {
     switch (action.type) {
       case Actions.GAME_EVENT: {
         state = { ...state, deltalog: [] };
@@ -295,13 +306,10 @@ export function CreateGameReducer({
         return state;
       }
 
+      case Actions.RESET:
       case Actions.UPDATE:
       case Actions.SYNC: {
         return action.state;
-      }
-
-      case Actions.RESET: {
-        return initial;
       }
 
       case Actions.UNDO: {