+++ docs/CHANGELOG.md
@@ -1,3 +1,33 @@
+## v0.30.0
+
+#### Features
+
+* [[6cf81e8](https://github.com/nicolodavis/boardgame.io/commit/6cf81e8)] create initial game state outside reducer
+* [[8d08381](https://github.com/nicolodavis/boardgame.io/commit/8d08381)] add a loading component for multiplayer clients
+
+#### Bugfixes
+
+* [[d20d26c](https://github.com/nicolodavis/boardgame.io/commit/d20d26c)] make master write to proper namepspaced keys
+
+## v0.29.5
+
+#### Features
+
+* [[7188222](https://github.com/nicolodavis/boardgame.io/commit/7188222)] add plugin.onPhaseBegin
+
+## v0.29.4
+
+#### Features
+
+* [[c1b4a03](https://github.com/nicolodavis/boardgame.io/commit/c1b4a03)] add playerSetup option to PluginPlayer
+
+## v0.29.3
+
+#### Features
+
+* [[da1eac6](https://github.com/nicolodavis/boardgame.io/commit/da1eac6)] rename plugin api functions
+* [[659007a](https://github.com/nicolodavis/boardgame.io/commit/659007a)] pass game object to plugins
+
 ## v0.29.2
 
 #### Bugfixes